+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+### ⚡ Changes
+
+- Add fallback to `PrimaryDirectoryPath` & `StorageVolumesDirectoryPaths` to be `"/storage/emulated/0"` & `["/storage/emulated/0"]` respectively.
+- Mark `MusicDirectoryPath` as deprecated mainly due to not specifying a fallback (ie: it may be `null`).
+  - Will be removed in `v1.0.0`.
+
+### 🛠️ Fixes
+
+- Attempt to fix `stacktrace: com.facebook.react.common.JavascriptException: Error: Exception in HostObject::get for prop 'MetadataRetriever': java.lang.NullPointerException: Attempt to invoke virtual method 'java.lang.String java.io.File.getAbsolutePath()' on a null object reference` that occurs on some devices.
+  - We don't have a device to replicate this issue so we're kind of going in blind.
+
 ## [0.6.0] - 2024-12-07
 
 > [!NOTE]